requirements.txt
streamlit>=1.28.0
python-nmap>=0.7.1
requests>=2.31.0
python-whois>=0.8.0
dnspython>=2.4.2
beautifulsoup4>=4.12.0
lxml>=4.9.0
pydantic>=2.0.0
urllib3>=2.0.0
pandas>=2.0.0